Enhancement: Extract DataProvider
