+++ test/Unit/Resource/RepositoryTest.php
@@ -28,7 +28,7 @@ final class RepositoryTest extends Framework\TestCase
     }
 
     /**
-     * @dataProvider providerInvalidOwner
+     * @dataProvider \Localheinz\GitHub\ChangeLog\Test\Util\DataProvider::providerInvalidRepositoryOwner
      *
      * @param string $owner
      */
@@ -50,17 +50,8 @@ final class RepositoryTest extends Framework\TestCase
         );
     }
 
-    public function providerInvalidOwner(): \Generator
-    {
-        foreach ($this->invalidOwners() as $key => $value) {
-            yield $key => [
-                $value,
-            ];
-        }
-    }
-
     /**
-     * @dataProvider providerInvalidName
+     * @dataProvider \Localheinz\GitHub\ChangeLog\Test\Util\DataProvider::providerInvalidRepositoryName
      *
      * @param string $name
      */
@@ -82,17 +73,8 @@ final class RepositoryTest extends Framework\TestCase
         );
     }
 
-    public function providerInvalidName(): \Generator
-    {
-        foreach ($this->invalidNames() as $key => $value) {
-            yield $key => [
-                $value,
-            ];
-        }
-    }
-
     /**
-     * @dataProvider providerValidOwnerAndName
+     * @dataProvider \Localheinz\GitHub\ChangeLog\Test\Util\DataProvider::providerValidRepositoryOwnerAndName
      *
      * @param string $owner
      * @param string $name
@@ -131,7 +113,7 @@ final class RepositoryTest extends Framework\TestCase
     }
 
     /**
-     * @dataProvider providerInvalidString
+     * @dataProvider \Localheinz\GitHub\ChangeLog\Test\Util\DataProvider::providerInvalidRepositoryString
      *
      * @param string $string
      */
@@ -146,41 +128,8 @@ final class RepositoryTest extends Framework\TestCase
         Resource\Repository::fromString($string);
     }
 
-    public function providerInvalidString(): \Generator
-    {
-        foreach ($this->invalidOwners() as $keyOne => $owner) {
-            foreach ($this->invalidNames() as $keyTwo => $name) {
-                $key = \sprintf(
-                    '%s/%s',
-                    $keyOne,
-                    $keyTwo
-                );
-
-                yield $key => [
-                    \sprintf(
-                        '%s/%s',
-                        $owner,
-                        $name
-                    ),
-                ];
-            }
-        }
-
-        foreach ($this->validOwners() as $key => $owner) {
-            yield $key => [
-                $owner,
-            ];
-        }
-
-        foreach ($this->validNames()  as $key => $name) {
-            yield $key => [
-                $name,
-            ];
-        }
-    }
-
     /**
-     * @dataProvider providerValidOwnerAndName
+     * @dataProvider \Localheinz\GitHub\ChangeLog\Test\Util\DataProvider::providerValidRepositoryOwnerAndName
      *
      * @param string $owner
      * @param string $name
@@ -198,159 +147,5 @@ final class RepositoryTest extends Framework\TestCase
         $this->assertInstanceOf(Resource\RepositoryInterface::class, $repository);
         $this->assertSame($owner, $repository->owner());
         $this->assertSame($name, $repository->name());
-    }
-
-    public function providerValidOwnerAndName(): \Generator
-    {
-        foreach ($this->validOwners() as $keyOne => $owner) {
-            foreach ($this->validNames() as $keyTwo => $name) {
-                $key = \sprintf(
-                    '%s/%s',
-                    $keyOne,
-                    $keyTwo
-                );
-
-                yield $key => [
-                    $owner,
-                    $name,
-                ];
-            }
-        }
-    }
-
-    /**
-     * @return string[]
-     */
-    private function invalidOwners(): array
-    {
-        $faker = $this->faker();
-
-        $values = [
-            'blank' => '  ',
-            'empty' => '',
-            'starts-with-hyphen' => \sprintf(
-                '-%s',
-                $faker->word
-            ),
-            'ends-with-hyphen' => \sprintf(
-                '%s-',
-                $faker->word
-            ),
-            'has-multiple-successive-hyphens' => \sprintf(
-                '%s--%s',
-                $faker->word,
-                $faker->word
-            ),
-            'has-underscores' => \sprintf(
-                '%s_%s',
-                $faker->word,
-                $faker->word
-            ),
-            'has-special-characters' => \implode('', $faker->randomElements([
-                '.',
-                '_',
-                ':',
-                'ä',
-                'ü',
-                'ö',
-                'ß',
-                '🤓',
-            ])),
-        ];
-
-        return $values;
-    }
-
-    /**
-     * @return string[]
-     */
-    private function validOwners(): array
-    {
-        $faker = $this->faker();
-
-        return [
-            'digit' => $faker->randomDigitNotNull,
-            'letter' => $faker->randomLetter,
-            'word' => $faker->word,
-            'word-with-numbers' => \sprintf(
-                '%s%d',
-                $faker->word,
-                $faker->numberBetween(1)
-            ),
-            'words-separated-by-hyphen' => \implode(
-                '-',
-                $faker->words()
-            ),
-            'words-with-numbers-separated-by-hyphens' => \implode(
-                '-',
-                \array_merge($faker->words(), [
-                    $faker->numberBetween(1),
-                    $faker->numberBetween(1),
-                ])
-            ),
-        ];
-    }
-
-    /**
-     * @return string[]
-     */
-    private function invalidNames(): array
-    {
-        return [
-            'blank' => '  ',
-            'empty' => '',
-            'has-special-characters' => \implode('', $this->faker()->randomElements([
-                '/',
-                '\\',
-                ':',
-                'ä',
-                'ü',
-                'ö',
-                'ß',
-                '🤓',
-            ])),
-        ];
-    }
-
-    private function validNames(): array
-    {
-        $faker = $this->faker();
-
-        return [
-            'digit' => $faker->randomDigitNotNull,
-            'letter' => $faker->randomLetter,
-            'word' => $faker->word,
-            'hyphen' => '-',
-            'hyphens' => '---',
-            'underscore' => '_',
-            'underscores' => '___',
-            'word-with-numbers' => \sprintf(
-                '%s%d',
-                $faker->word,
-                $faker->numberBetween(1)
-            ),
-            'words-separated-by-hyphen' => \implode(
-                '-',
-                $faker->words()
-            ),
-            'words-separated-by-underscore' => \implode(
-                '-',
-                $faker->words()
-            ),
-            'words-with-numbers-separated-by-hyphens' => \implode(
-                '-',
-                \array_merge($faker->words(), [
-                    $faker->numberBetween(1),
-                    $faker->numberBetween(1),
-                ])
-            ),
-            'words-with-numbers-separated-by-underscores' => \implode(
-                '_',
-                \array_merge($faker->words(), [
-                    $faker->numberBetween(1),
-                    $faker->numberBetween(1),
-                ])
-            ),
-        ];
     }
 }
